+++ ext/stub/properties/propertyarray.zep.c
@@ -64,7 +64,7 @@ PHP_METHOD(Stub_Properties_PropertyArray, __construct)
 	ZEPHIR_INIT_VAR(&_1);
 	ZVAL_LONG(&_1, 1);
 	zephir_array_fast_append(&_0, &_1);
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 94, &_0);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 95, &_0);
 	ZEPHIR_MM_RESTORE();
 }
 
@@ -103,7 +103,7 @@ PHP_METHOD(Stub_Properties_PropertyArray, setOtherArray)
 		Z_PARAM_ZVAL(arr)
 	ZEND_PARSE_PARAMETERS_END();
 	zephir_fetch_params_without_memory_grow(1, 0, &arr);
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 95, arr);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 96, arr);
 }
 
 PHP_METHOD(Stub_Properties_PropertyArray, getOtherArray)
@@ -147,7 +147,7 @@ PHP_METHOD(Stub_Properties_PropertyArray, testIssues1831)
 		zephir_check_call_status();
 		ZEPHIR_INIT_NVAR(&_3$$3);
 		ZVAL_STRING(&_3$$3, "header");
-		ZEPHIR_CALL_FUNCTION(&_4$$3, "stripos", &_5, 94, &info, &_3$$3);
+		ZEPHIR_CALL_FUNCTION(&_4$$3, "stripos", &_5, 99, &info, &_3$$3);
 		zephir_check_call_status();
 		if (!ZEPHIR_IS_FALSE_IDENTICAL(&_4$$3)) {
 			zephir_array_append(&headers, &info, PH_SEPARATE, "stub/properties/propertyarray.zep", 51);
@@ -156,7 +156,7 @@ PHP_METHOD(Stub_Properties_PropertyArray, testIssues1831)
 			ZEPHIR_OBS_NVAR(&_6$$5);
 			_7$$5 = zephir_fetch_property_write(this_ptr, _zephir_prop_0, &_6$$5);
 			ZEPHIR_MAKE_WRITE_REF(_7$$5);
-			ZEPHIR_CALL_FUNCTION(NULL, "array_unshift", &_8, 95, _7$$5, &info);
+			ZEPHIR_CALL_FUNCTION(NULL, "array_unshift", &_8, 100, _7$$5, &info);
 			ZEPHIR_UNREF_WRITE(_7$$5);
 			zephir_check_call_status();
 			break;